+++ ts/src/encoding/mltEncoderSynthetic.spec.ts
@@ -2,9 +2,7 @@ import { readFile } from "node:fs/promises";
 import { describe, expect, it } from "vitest";
 import { compareWithTolerance, getTestCases } from "../../../test/synthetic/synthetic-test-utils";
 import decodeTile from "../mltDecoder";
-import { GEOMETRY_TYPE } from "../vector/geometry/geometryType";
-import { classifyRings } from "../vector/geometry/classifyRings";
-import type { Geometry } from "../vector/geometry/geometryVector";
+import { featureTablesToFeatureCollection, getGeometry } from "../vector/featureTablesToGeoJson";
 import type FeatureTable from "../vector/featureTable";
 import { encodeTile, type Feature, type FeatureGeometry, type Layer, type PropertyValue } from "./mltEncoder";
 
@@ -64,7 +62,7 @@ describe("encodeTile - synthetic fixtures round trip", () => {
 /** Reads a reference tile, re-encodes what it decodes to, and decodes that back to GeoJSON. */
 async function reEncode(mltFile: string): Promise<GeoJSON.FeatureCollection> {
     const reference = decodeTile(new Uint8Array(await readFile(mltFile)), undefined, false);
-    return toFeatureCollection(decodeTile(encodeTile(toLayers(reference)), undefined, false));
+    return featureTablesToFeatureCollection(decodeTile(encodeTile(toLayers(reference)), undefined, false));
 }
 
 /** Turns decoded feature tables back into the layer shape the encoder takes. */
@@ -83,27 +81,6 @@ function toLayers(featureTables: FeatureTable[]): Layer[] {
     }));
 }
 
-/** The same conversion the decoder's synthetic harness uses, so both sides are comparable. */
-function toFeatureCollection(featureTables: FeatureTable[]): GeoJSON.FeatureCollection {
-    const features: GeoJSON.Feature[] = [];
-    for (const table of featureTables) {
-        for (const feature of table.getFeatures()) {
-            const geojsonFeature: GeoJSON.Feature = {
-                type: "Feature",
-                geometry: getGeometry(feature.geometry),
-                properties: {
-                    _layer: table.name,
-                    _extent: table.extent,
-                    ...Object.fromEntries(Object.entries(feature.properties).map(([k, v]) => [k, safeNumber(v)])),
-                },
-            };
-            const id = safeNumber(feature.id);
-            if (id !== null && id !== undefined) geojsonFeature.id = id;
-            features.push(geojsonFeature);
-        }
-    }
-    return { type: "FeatureCollection", features };
-}
 
 /** Property order and key order are not preserved by encoding, so compare a canonical form. */
 function normalise(collection: GeoJSON.FeatureCollection): GeoJSON.FeatureCollection {
@@ -121,28 +98,4 @@ function normalise(collection: GeoJSON.FeatureCollection): GeoJSON.FeatureCollec
             return normalised;
         }),
     };
-}
-
-function safeNumber<T>(val: bigint | T): T | number {
-    return typeof val === "bigint" ? Number(val) : val;
-}
-
-function getGeometry(geometry: Geometry): GeoJSON.Geometry {
-    const coords = geometry.coordinates.map((ring) => ring.map((p) => [p.x, p.y]));
-    switch (geometry.type) {
-        case GEOMETRY_TYPE.POINT:
-            return { type: "Point", coordinates: coords[0][0] };
-        case GEOMETRY_TYPE.LINESTRING:
-            return { type: "LineString", coordinates: coords[0] };
-        case GEOMETRY_TYPE.POLYGON:
-            return { type: "Polygon", coordinates: coords };
-        case GEOMETRY_TYPE.MULTIPOINT:
-            return { type: "MultiPoint", coordinates: coords.map((r) => r[0]) };
-        case GEOMETRY_TYPE.MULTILINESTRING:
-            return { type: "MultiLineString", coordinates: coords };
-        case GEOMETRY_TYPE.MULTIPOLYGON:
-            return { type: "MultiPolygon", coordinates: classifyRings(coords) };
-        default:
-            throw new Error(`Unsupported geometry type: ${geometry.type}`);
-    }
 }
